+++ parser/client.py
@@ -5,54 +5,14 @@ from requests.adapters import HTTPAdapter
 from requests.utils import dict_from_cookiejar
 from requests_cache import CachedSession
 from http.cookiejar import MozillaCookieJar, CookieJar
-from furl import furl
 from fake_useragent import UserAgent
-from typing import Final, Optional, Union, Mapping, FrozenSet
-from dataclasses import dataclass, field
+from typing import Final, Optional, Union, Mapping
 from abc import ABC
 from os import PathLike
 from pathlib import Path
+from configs import ClientConfig, NetConfig, CacheConfig, CookieConfig
 
 SessionLike = Union[NotCachedSession, CachedSession]
-
-
-@dataclass(slots=True)
-class ClientConfig:
-    base_url: furl
-    name: str
-
-    def __post_init__(self):
-        self.base_url = furl(self.base_url).remove(fragment=True, args=True)
-
-
-@dataclass(slots=True)
-class NetConfig:
-    timeout: float = 5.0
-    retries_total: int = 3
-    retries_connect: Optional[int] = None   # if None retries_connect = retries_total
-    retries_read: Optional[int] = None      # if None retries_read = retries_total
-    backoff_factor: float = 0.5
-    allowed_methods: tuple[str, ...] = ("GET", "HEAD")
-    status_forcelist: FrozenSet[int] = frozenset({429, 500, 502, 503, 504})
-
-    headers: dict[str, str] = field(default_factory=dict)
-    proxies: dict[str, str] = field(default_factory=dict)  # {"http": "...", "https": "..."}
-    user_agent: Optional[str] = None  # if None use random user agent
-
-
-@dataclass(slots=True)
-class CacheConfig:
-    enabled: bool = False
-    ttl: int = -1  # Immortal cache
-    name: Optional[str] = None
-    cache_control: bool = False  # Don't respect server cache
-    backend: str = "sqlite"
-
-
-@dataclass(slots=True)
-class CookieConfig:
-    initial: Optional[Union[Mapping[str, str], CookieJar]] = None
-    file: Optional[Union[PathLike[str], str]] = None
 
 
 class Client(ABC):
